Adds flag for resetOnStop
This flag will prevent the animation from resetting when the animation is complete.  This is useful because it allows you to use the OneShotAnimation Controller to create a splash screen and to be able to tie into the onStop callback to close the splash screen or toggle a flag letting you know that the splash screen has completed, and that you can dismiss the splash screen when your data has finished loading.