chore: end with call to dispose
Fixes: https://github.com/rive-app/rive-flutter/issues/335

By calling `super.dispose()` at the end - as suggested in the Flutter docs for widgets + render objects.

Added the same to the editor (for widgets + render objects)

These changes _should_ not result in any issues. But I don't know if there are valid reasons for any of these to call super.dispose first - if yes, then we should investigate the underlying reason for that.

Diffs=
913760d97 chore: end with call to dispose (#5703)

Co-authored-by: Gordon <pggordonhayes@gmail.com>