Fix loading init text in Flutter runtime.
The important work is done in _initTextAndImport. This functionality needs to be called by the user if they call RiveFile.import. See the changelog for some notes.

I needed to bump rive_common for this to work as we manually added our fixRequireJs fix to the HTML. We should always try to inject functionality that's required at runtime dynamically so it can apply to all apps (not just our editor).

Diffs=
d072cbde0 Fix loading init text in Flutter runtime. (#5758)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>