Improve collapse propagation for solo.
Fixes https://2dimensions.slack.com/archives/CLLCU09T6/p1692895941301969

Should also address #5888

Basically we were propagating the collapse to the ClippingShape for the solo which meant that the update on the clip was never called. This meant the path would never be generated and so our contents would always get clipped out (no intersection with an empty path).

ClippingShapes and Constraints should be treated more like siblings (or properties) of the Solo instead of children that can be collapsed by the active solo child choice.

Diffs=
e1b9d360b Improve collapse propagation for solo. (#5890)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>