Core.copy doesn't copy core object id.
Usually when we copy/clone an object we do it with the intention of adding to core shortly after. This has a problem if the object has an id already, it won't be assigned a new one which means we'll end up with multiple objects in core with the same id. This breaks the fundamental concept in Core that each Id will always point to an object of a specific type (which is important for undo/redo but also for concurrent client changes). Changing a type should incur a new core object.

Note that reference ids to other objects still get copied.

Diffs=
f9a356004 Core.copy doesn't copy core object id. (#6187)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>