ignore paths that are inactive in solos when calculating hit test
we were not skipping collapsed shapes and paths for hit testing, so pointer events were triggering on elements that were inactive.
This PR skips those paths to fix it.
Note: there might be some performance improvements that can be done in the future skipping shapes that are fully hidden, but this PR most likely covers the majority of the usual cases.

Diffs=
2c2d332e0 ignore paths that are inactive in solos when calculating hit test (#6276)
c37a28468 generate drawing rules in the correct order (#6275)

Co-authored-by: hernan <hernan@rive.app>