fix for unexpected triggered events
fixes #7226
This PR fixes reports of events when animations are playing backward.
It considers both the timeline speed and the state machine state speed to determine direction.
It also handles not reporting an event twice if it comes from a "ping pong" loop.
It also adds a small performance improvement, skipping the binary search if possible.

Diffs=
265c00985 fix for unexpected triggered events (#7227)
8a538c243 validating core objects property keys on load (#7298)
cd6210f42 Make ContourMeasure more robust (#7294)
405b8ef90 use world bounds for coarse grained collision test (#7287)
3734d9bac initialize seed with chrono (#7285)
45b998e68 Simple libjpeg (#7277)

Co-authored-by: hernan <hernan@rive.app>