Xxxx fix extra frame rendering
this PR fixes two sources of neverending renders.
the first commit addresses a scenario where, once the stage is zoomed in or out, it gets into a continuous scheduling of frames that prevent the cache from being rebuilt because the debounce resets the previous one. It is fixed by storing a pending zoom value, and only resetting the debounce if the pending value is different than the new zoom value.

The second commit fixes a scenario where, when time remapped animations are advanced in preview mode, it also triggers the advancement of the nested artboards belonging to the remapped artboards.

Diffs=
7474216b4 Xxxx fix extra frame rendering (#7301)

Co-authored-by: hernan <hernan@rive.app>