Nested linear animations report events up to parent artboards
Previously, only nested state machines could report events so that listeners in parent artboards could listen for them. This PR adds event reporting for nested simple animations. Had to refactor some stuff to genericize in order for both state machines and linear animations to have similar functionality.

I'm not sure if its possible, or desirable, for nested remap animations to have the same functionality, but that is not included in this PR.

Diffs=
097b68f56 Nested linear animations report events up to parent artboards (#7310)
34e186b32 more renames for harfbuzz (#7398)
b88272290 mark shape as dirty after flagged as target (#7396)
a10b1e61e don’t defer updates when a shape/path is used for hit detect (#7392)

Co-authored-by: Philip Chung <philterdesign@gmail.com>