Xxxx improve hittest performance
In certain scenarios, there is no need to perform a hit test, so we precompute the conditions and early out from calculating the hit test.
If a shape has only listeners of type PointerDown and PointerUp, and is not an opaque target, it doesn't need to check for move events or exit events, which can save a lot of computations since it will skip most frames.

Diffs=
50bc398c4 Xxxx improve hittest performance (#7584)
a01b0467e Makeshadersinpremake (#7660)
114da4e39 Layout drawable (#7544)
b5f342002 add bindable properties for state machines (#7640)
f55ebd34f skip custom events when creating hitshapes (#7577)

Co-authored-by: hernan <hernan@rive.app>