add support for state machine conditions based on view models
This PR defines the core objects needed to support adding conditions to transitions based on view model values.
It also adds the editor UI necessary to create them based.
I added some comments on the most relevant parts of this PR.
The c++ part will be added on a separate PR to keep this one focused on the pieces needed to create these conditions.
The rest is boilerplate code from generate_core and generate_core_runtime.

Diffs=
190b920f8 add support for state machine conditions based on view models (#7659)
1a5f273bb Nested artboard types: node, leaf, layout (#7639)

Co-authored-by: hernan <hernan@rive.app>