feat(flutter): add multitouch support (#10688) d5e74de08c
fix(vulkan): Correctly support Vulkan 1.0 on Android (#10679) 8e0fadf978
VkBootstrap will incorrecty error with `vulkan_version_1_1_unavailable` on Vulkan 1.0 systems that do not have the vkEnumerateInstanceVersion function, even when we specifically said that it was fine to fall back to 1.0. To work around this, we will detect that error and then try again to create the Vulkan instance, specifically requesting version 1.0.

Additionally fix a "function must return" error when building in MSVC on Windows, and update LOG_ERROR to be LOG_ERROR_LINE to make it clear that it writes lines (and update its use of fprintf to match what the android logging does)

chore: Rename draw_clockwise_* shaders to draw_clockwise_atomic_* (#10689) 7e676124de
The interlock mode these shaders implement is clockwiseAtomic. Let's be
more specific with the naming so we can support other clockwise modes.

Co-authored-by: Gordon <pggordonhayes@gmail.com>